Importing scripting bridge
The scripting bridge is based on manually written Pyrex based wrappers over the
exposed application APIs. Initial support is here added for reading and writing
config variables and connecting scripts to I/O events. In the future, the
object manager will have a script wrapper as well, but this functionality is at
the moment missing..


git-svn-id: https://openorbit.googlecode.com/svn/trunk@13 67975e55-dd32-0410-8bc3-a9813c38a78a