Initial resource manager functions for MacOS X
The resource manager will locate and open resource files used by the
application. The files in this checkin are Mac specific, allowing us to bundle
resources in the application bundle.

Ideally these should be implemented for generic POSIX and WIN32 as well. For the
moment this is down prioritised as the main goal at the moment is to do
sufficient clean-up in order to do a first public release of the source code.


git-svn-id: https://openorbit.googlecode.com/svn/trunk@46 67975e55-dd32-0410-8bc3-a9813c38a78a