Adding stars loaded from a file
Open Orbit now loads stars from a resource file. The resource locating
functions have been exposed to python in the res.pyx file, and the
environment.pyx that is intended to map functions for setting up the
natural environment (inserting stars, planets, etc etc) have been added.

A file containing all the stars of magnetude 6.0 or less have been added.
The star database contain data from the hipparcos database, but heavily
stripped down. The hip filter script can be used to regenerate the file
from the original hipparcos database available from ESA. You have to
view the script source in order to figure out how to use it at the moment.
The data used by open orbit is vmag, dec, ra and b-v colour (that is
converted with a function to blackbody temp and then looked up in a clut).
The function for converting b-v to blackbody is attributed to the paper on
star colours written by Thor Olson, available at:
http://www.nightscapes.net/techniques/TechnicalPapers/StarColors.pdf

Note that all stars are drawn as single pixels with their magnitude
roughly converted to an alpha value, this mechanism is subject to change
but should at least give a somewhat decent result.


git-svn-id: https://openorbit.googlecode.com/svn/trunk@69 67975e55-dd32-0410-8bc3-a9813c38a78a