Initial support for proxy objects in the object manager
The proxy objects are objects in the object manager that allows the user
to explicitly control the allocation and deallocation of the objects.

Update adds suppor to alloctate proxy classes, where all objects of a
certain class are assumed to be manually managed (i.e. no calls to create
or destroy).


git-svn-id: https://openorbit.googlecode.com/svn/trunk@70 67975e55-dd32-0410-8bc3-a9813c38a78a