Hashtable now keeps a list of all the hash table entries
The hashtable is now more or less a hybrid list/dictionary. The underlying
mechanism is there, but it is not yet exposed on a higher level. This
functionality will be needed by the object manager (as it will be able to
dump the dual maintainance of lists and hashtables). The hashentries are
now pointing at the list entries and the other way around, the missing
functionality (such as possibility to iterate over the table) will be
added shortly.


git-svn-id: https://openorbit.googlecode.com/svn/trunk@71 67975e55-dd32-0410-8bc3-a9813c38a78a