Dynamic indexing in om properties should now allow getting of values
The array indexing now allows for dynamic arrays to be read, it will check
the array bounds and if failed print an error message and exit. The exit
functionality is not really good to have here, but will remain for the time
being. All the abort calls have been replaced with calls to errx, which
should at least generate some kind of usual information on failure.


git-svn-id: https://openorbit.googlecode.com/svn/trunk@77 67975e55-dd32-0410-8bc3-a9813c38a78a