Build system now installs resources in Mac bundle
The build system have been updated to install resources in the application
bundle when building on a Mac. The placement of the installation commands
might not be the best in simplicity point of view, we should probably have
a CMakeLists.txt in the rsrc directory instead, but it is good enough for
now.


git-svn-id: https://openorbit.googlecode.com/svn/trunk@79 67975e55-dd32-0410-8bc3-a9813c38a78a