Xcode project file and minor fixes
Xcode project now reflects the actual files in the project. Minor clanup of
whitespace and adding the __init__.py file that will be in the future
python package. Note that the pyrex files does not support the package
layout yet (and the static linking might need to be replaced with a shared
lib).


git-svn-id: https://openorbit.googlecode.com/svn/trunk@90 67975e55-dd32-0410-8bc3-a9813c38a78a