Generalisation of vectorisation code and default build type
The altivec specific types have been changed to use the general gcc
attribute system for defining vector types. This should work for x86 SSEx
extensions as well as for altivec. Have moved the setting of compile flags
to a the file cmake/arch-flags.cmake, this is intended to make it easier
to maintain arch-specific optimisations. The build type will now default
to Release if nothing is specified.

git-svn-id: https://openorbit.googlecode.com/svn/trunk@93 67975e55-dd32-0410-8bc3-a9813c38a78a