Remove macro interface to vmath
The macro interface to vmath is now no longer
supported a new function based interface is in
development. This new interface will support
optimisations through inline functions. It can be
noted here however that he inline interface is not
in place at the moment, and more thoughts need to
be put into it.


git-svn-id: https://openorbit.googlecode.com/svn/trunk@117 67975e55-dd32-0410-8bc3-a9813c38a78a