Make sure sg draw function walks through all the nodes
The sg draw procedure is now split up into two procedures. One that is doing
fundamental settings for the entire scene (e.g. setting the current matrix to
MODELVIEW and loading the identity), and the static sub function sg draw 2
that is drawing all the nodes by recursively invoking itself on child nodes.