Clean-up scripting init so that no result is returned
Init function will now exit on errors. The rationale for this is that all sub-
systems must successfully initialise themselfes before we can do anything. Thus
all failures in the sub system inits are fatal anyway, so we might as well have
a clean main and call exit on error.

git-svn-id: https://openorbit.googlecode.com/svn/trunk@150 67975e55-dd32-0410-8bc3-a9813c38a78a