Initial work on scenegraph compliant camera
The new camera system will be fully integrated in the scenegraph. The support
added here flesh out the initial support for this. It is now also the
responsibility of the postinit script to initialise a camera and make it the
default camera.

This also include a python wrapper for the mesh objects.