Add better scenegraph structures and create files for GUI structures
The scenegraph should be more flexible and this requires that we can start to
draw in the node where the camera is placed. Thus the nodes must know their
parents for making the graph traversal simple.

GUI files will contain structures used for defining menus and other GUI overlay
elements.