Add geometric functions for creating ellipses
The new ellipse structure and functions supports creating ellipses with equal
area segments (with respect to one of the foci). This is tailored to support the
pre-evaluation of simple orbits and the caching of these orbits in memory. In
term, this will let us reduce the number of objects that are involved in n-body
calculations.