Add ellipse point interpolation function
The function can be used to interpolate a point between two known points on the
ellipse edge. The interpolation is linear, but should be rougly accurate.

This is especially useful in order to estimate the orbital point of an object
in a perfect elliptical orbit. The estimation would then be done by calling the
function passing in the time t scaled so that one year / orbit in t is expressed
in the number of segments in the ellipse (as defined when creating it). Thus, if
the number of segments is 100, and t is 50.3, you will get a value between point
50 and point 51, interpolated with 0.3, that is p[50] + (p[51]-p[50]) * 0.3.