Add initial code for autowrapper tool
The autowrapper tool is intended to be used for wrapping C-code with the python
API. The goal of this is to in the end replace the pyrex code with this code
that will eliminate the need to handcraft wrappers in pyrex (which is a very
error prone and time consuming process). The autowrap tool is based on the clang
C compiler and is written in C++(!), yes... that is correct, C++.

Though, while this project (openorbit) normally avoids C++ like fire, it is
justified in this case as it allows us to use a fully working C compiler
frontend. In the long run, autowrap may be extended to support other target
languages but python and to support both Objective-C and C++ as source
languages.