Autowrap support for multiple input files
Autowrap now supports multiple input files passed on the command line. It also
adds some additional logic for identifying functions that shuould be added as
instance or class methods when being wrapped.

In the long run, this behaviour should probably be specified with attributes
to the functions or pragmas. However, for now, we will be satisfied with
automatic handling of these cases (i.e. when the first parameter is a record
pointer or when the result is a record pointer).