Make plugin loading work
Plugins now load and resolve backlinks. The cmakefiles updates are a bit mac
speciffic but the commit serves as a starting point.

Problems when loading plugins are that we do not want to have the openorbit exec
divided into multiple libraries, so the gencds lib is now simply built into
the openorbit target. This allows gencds symbols to be resolved properly when
loading a plugin.