Human Readable Markup Language Support
HRML has been defined for this project, the HRML language is a human readable
markup language, it is intended to be compatible with XML. But unlike XML,
editable with a standard text editor.

It is more sophisticated than XML as it supports types (e.g. strings, floats,
integers et.c.).

Note that the HRML parser is not yet finished, and some additional work is
needed. The parser is written in C in order to eliminate dependencies on tools
like YACC/Bison et.c. The implementation here is the first HRML implementation,
and in the future, a more formal HRML grammar will be published including the
reference parser. At the moment, the encoding support is also limited, but it is
the intention that all documents that are HRML compliant must be saved in
UTF-{8|16|32} format and that other encodings are strictly forbidden.

One way to extend the system to UTF support is to allow the strings to contain
UTF sequences, since the lexer is handwritten and does not depend on FLEX this
should be a relatively trivial task.