Update tests for de-libified libgencds
The generic C data structure library has seased to be a library in order to ensure that
it is exported and linked into the main application which will provide the CDS lib as
exported APIs to the plugins. If it is a library the darwin linker will strip unused
objects from it. When this was fixed, the tests where not changed accordingly, this is
hereby addressed.