Centralize scripting API
Pyrex was causing to much problems when distributing the
scripting support in multiple files. The longterm goal is
anyway to get rid of Pyrex and have a clang-based python
wrapper generator. This commit moves all the scripting
support to a central pyrex module named ooscript. This
is obviously far from ideal, and the module look like crap
at the moment, but at least it works.

Currently openorbit craches when it try to load an hrml
file with solar system data.