New hrml parser getting ready
The new parser now have more functionality than the original
parser. The original parser is subsequently removed and the
new parser has been graduated as the primary one.

Currently missing functionality is dates, arrays, anonymous
members and units. Units are supported on parse level, but
ignored and returned as a float.