Relicensing to LGPLv3
In order to reduce confusion, the MPL will no longer be applicable to Open Orbit. Open Orbit was previously
triple licensed under the MPL, GPL and LGPL. It has now
moved to the LGPLv3 permanently.

This will effectively protect Open Orbit while still
allowing plug-ins to be distributed in whatever licence
that the plugin-developer sees fit.