Move Camera Rotation Logic to Camera Source
Previously, the SG paint function was responsible for camera
transforms and rotations, this is now handled by the camera
subsystem itself.

Two functions exist, one for applying only rotation, and one
for applying camera rotation and translation. This allows the
SG to apply a rotation before drawing the sky, but ignore the
translation in that case.