Space Craft System Simulation
Moving forward with spacecraft system simulation. The idea is
now that each spacecraft has a vector of stages, and each stage
has a vector of engines and torquers. Engines may be enabled in
multiple stages (e.g. space shuttle), the active stages are
tagged with an enabled flag, performance wise, scanning each
stage is not an issue since a spacecraft would typically have
very few stages.