Move large world coordinates to floats
This is unfortunately necessary for compatibility with the
rendering pipeline. This also means that we have to increase
the integer components for the sectors to 32 bit, this in term
allow us to work with regions of around 20 au in size.
Though this is still not enough to represent the entire solar
system, the solar system in itself will probably consist of
multiple disjoint regions of these segmented spaces.