Add support for double based segments
This might need to be used in the physics engine, but this is
unknown at the moment. The important thing in that case is that
it is easy to switch from floats to doubles and vice versa.
Even though the segment lengths of a double based segment could
be a lot larger, they must unfortunately be restricted to the
valid float ranges (around 1024.0), as the data needs to be
moved to the rendering pipeline at some point.