Prepare for major physics system / scenegraph surgery
This includes some cleanup. Firstly, some previously public
data structures in the scenegraph system has been made private.
Secondly, the pushing of data to the scenegraph is now done by
the physics system and not by scenegraph supplied functions.
The physics system also use accessors when pushing object data
into the scenegraph. This will in term simplify the
introduction of large world coordinates in the physics system
and the use of localised coordinates in the scenegraph.