Planetary rendering
Allthough this is just a copy of the sphere rendering at the
moment. We do in the long run need specialised renerers for the
planets. Probably one for rocky planets and one for gas giants.

In any case, this keeps the planet renderer up to date with the
rest of the scenegraph.