Default endianess assumption
Normally, the endianess should be tested by a configure
script. But for the bitutils, default assumptions based
on preprocessing macros is reasonable. Note that, the
processor tags need to be verified, I only entered them
from memory, they may be wrong, but I only have an x86
at the moment.