Working Xcode project
The Xcode project is now working. Currently, installing
resources in the CMake system is experiencing some trobubles
but this will be fixed as soon as possible.

Note that the xcconfig file keeps track of various locations
for headers and so on that can easily be customized to your
local system.