Fix bad first segment in orbits and polish up the orbital drawing
Orbits are now drawn by more simple ellipses than the ones used
to calculate the orbital position. Reason for this is that the
orbital position ellipse need more segments as it computes
segments based on area sweeps. For small eccentricity values,
this is not a problem, but when the ecc goes up and the orbit is
elongated a higher number of area segments will be needed. For
drawing however, we can use ellipses of fewer but more uniform
segments.

In any case, drawing orbits now actually look OK.