Object and camera rotation
Firstly, fixes issue where the rotation matrix for
objects need to be transposed as gl wants column
major ordered matrices. This does not apply to the
camera as the camera is rotating the world instead
(which ends up being the inverse of what we think
we are doing in rotating the camera).