AC3D parser works!!!
The ac3d file loader now successfully parses an
AC3D file. Though, it may not be complete and
fully working, and it certainly miss a lot of
validation checks for ensuring that files are
correct with respect to the AC3D file format.

Still to fix are the conversion of the file format
graph into a more suitable internal format for
rendering.