Add missing scenegraph files
Note, these are at the moment place holders and they actually
do not really do anything at the moment.

Terrain files are supposed to hold the implementation of a
lod based terrain renderer.

Text files are supposed to hold the implementation of a
screen-printf like function that allow you to print text
at a certain position on the screen.