Mercury thruster postions
The roll, pitch and yaw thruster positions should
be roughly correct now. This data is from
measuring the interior arrangement plans. Origin
for positions is the center of the back plate of
the capsule. This means that the retro pack is
still incorrectly placed, but it should be its own
stage anyhow.