Code cleanup plus additional field for PLobject
Nothing really new, mostly cleanup. But, adding an additional field to PLobject to keep track
of parent objects. Parent objects are objects that have children, but is treated as a unit.

At later stage, however a child object may be detatched from the parent and the child is then a
free object and the parent will get its mass parameters modified.