Working orbiting camera
The orbiting camera is now working (some bugs with
ra angles). Start-up-script now centers camera on mercury
spacecraft in orbit around earth. This allow us to start
implementing and testing the spacecraft systems.

Also include a bugfix for star-rendering, it should
render with lights off, but this was not explicit.