Smooth orbital simulation
Due to rounding errors, sometimes the orbital simulation
did not converge good enough. Now the orbital velocity
is evaluated and the system does newton steps based on them
every couple of seconds, the orbital position is updated by
reevaluating the solver for it and setting a new velocity
vector. This smothes out some jumpiness that occurred due
to the non converging solver.