Spacecraft plugin system and Mercury spacecraft
The spacecraft plugin system is extended to allow static
plugins, this replace the previous HRML loader since
there are plenty of simulation aspects that are turing complete.

The mercury hrml file has been ported to the new system
and several extensions to the simulation system has been
added and implemented, allowing for the ignition of the
main engine and the detatchment of stages.