Add preliminary text rendering code
Introduces some simple functions for rendering text into grayscale bitmaps. At present, it has been somewhat tested in a small test
application, but is rather messy and buggy. It will for example not
realloc the image buffer if it is not in the right size.

Note that this commit introduce a dependency on the freetype package.
For the Xcode build files, the dependency is on the Mac Ports
installation of freetype.