Planet shaders
Planet shader now allows for the use of a night texture that will
be used instead of the ambient color (for the terminator/twilight zone
there is blending between the day texture (with specular reflections)
and the night texture. Further, the spheres are now loaded with a
material (used when no texture is active and for specular hilights). The specular is also adjusted using a specular map, that can be used
to enable / disable regions on the planet for specular lighting (e.g.
water should reflect, but ground should not).