Massive I/O system rework (gets rid of SDL on OS X)
This mega commit gets rid of SDL requirements on OS X. It replaces
the last dependency on SDLs joystick code with direct access to
the OS X HID manager APIs. Note, that this code needs to be ported
to Linux/BSD/Windows.

As a bonus, joystick support now supports hot plugging (and
unplugging) of devices.