Rewrite pubsub API
This rewrite does a massive cleanup of the pubsub code. Especially
it provides new types for all basic C-types that should be used
when declaring new variables if they should be visible in the
pubsub system.

Also, the axises from the I/O-manager are now routed through the
pubsub system.