Changing license of Open Orbit
Most of the code is now licensed under the GPLv3 instead of the previous LGPLv3.
Some files have however been relaxed and are now covered by the modified BSD
license, this is an exception that is granted on a file by file basis.

Essentially, the modified BSD license is applied on code that is general in
nature and not directly related to the simulation code as such. For example, the
generic data structure code providing dynamic arrays, linked lists, AVL trees,
heaps and hashtables is now under the modified BSD license. Again, check the
comment at the start of the files, the files typically say something in the
line of: "This file is licensed under the 2-clause BSD license" or similar.