Module initialization order fixes
According to hidden GCC documentation, the constructor priority
is only allowed to be set to 101 and higher (max 64k). This is
hereby addressed. However, at present a clang bug will result in
the constructors being executed in the wrong order, if located in
separate compilation units.

Unfortunately, this means that openorbit builds are currently
broken :(